+++ CRM/Civirules/Upgrader.php
@@ -610,6 +610,7 @@ WHERE contact_id NOT IN (select id from civicrm_contact c where c.id=rl.contact_
 
   public function upgrade_3001() {
     CRM_Civirules_Utils_Upgrader::insertTriggersFromJson(E::path('sql/triggers.json'));
+    CRM_Civirules_Utils_Upgrader::insertConditionsFromJson(E::path('sql/conditions.json'));
     return TRUE;
   }